GVA: Use RecyclerView for the Persistent button options
The ButtonAdapter from the GvaGalleryItemViewHolder was extracted to a separate file. Now it is used for both GvaGalleryItemViewHolder and GvaPersistentButtonsViewHolder.

MOB-2520
