+++ widgetssdk/src/main/res/layout/chat_gva_persistent_buttons_content.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
     android:layout_marginTop="@dimen/glia_medium"
     android:background="@drawable/bg_message">
@@ -20,10 +20,12 @@
         android:textIsSelectable="true"
         android:textColorLink="?attr/gliaBaseDarkColor" />
 
-    <LinearLayout
-        android:id="@+id/container"
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/buttons_recycler_view"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical" />
+        android:paddingVertical="@dimen/glia_medium"
+        android:paddingHorizontal="@dimen/glia_large"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
 
 </LinearLayout>
